Fix benchmarks + more
This commit fixes a major flaw in benchmarks. The state was initialized
only for the first iteration, not for all of them. This means that only
the first iteration worked as intended, other iterations have worked on
already sorted data.

This commit also adds a .clang-format file and it formats all code.